Rename AbstractErrorHandler to ErrorHandler
